# internal/api/open_api/OpenAPISpec.gen.yaml
openapi: 3.1.0
info:
  title: Bridgr API
  description: Skill-gap navigator HTTP API (decoupled from hassleskip/users legacyapi).
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: http://api:8080
paths:
...
  /v1/bridgr/graphs/{graphUUID}/nodes:
    get:
      summary: List nodes for graph
      operationId: V1GetBridgrGraphNodes
      tags:
        - bridgr
      parameters:
        - name: X-API-KEY
          in: header
          description: API Key of the service making the request
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: graphUUID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgrSkillGapNodeListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Graph not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create graph node
      operationId: V1PostBridgrGraphNodes
      tags:
        - bridgr
      parameters:
        - name: X-API-KEY
          in: header
          description: API Key of the service making the request
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: graphUUID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBridgrSkillGapNodeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgrSkillGapNode'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Graph not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/bridgr/graphs/{graphUUID}/nodes/by-key/{nodeKey}:
    get:
      summary: Get node by graph and node_key
      operationId: V1GetBridgrGraphNodeByKey
      tags:
        - bridgr
      parameters:
        - name: X-API-KEY
          in: header
          description: API Key of the service making the request
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: graphUUID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: nodeKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgrSkillGapNode'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/bridgr/graphs/{graphUUID}/edges:
    get:
      summary: List edges for graph
      operationId: V1GetBridgrGraphEdges
      tags:
        - bridgr
      parameters:
        - name: X-API-KEY
          in: header
          description: API Key of the service making the request
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: graphUUID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Edges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgrSkillGapEdgeListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Graph not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create directed edge in graph
      operationId: V1PostBridgrGraphEdges
      tags:
        - bridgr
      parameters:
        - name: X-API-KEY
          in: header
          description: API Key of the service making the request
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: graphUUID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBridgrSkillGapEdgeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgrSkillGapEdge'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Graph not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
...
    NotFoundErrorResponse:
      type: object
      required:
        - error
        - not_found_by_api
      properties:
        error:
          type: string
          description: Error message
        not_found_by_api:
          type: boolean
          description: |
            Always true. This key indicates that the API successfully fielded the request and that the specific resource requested
            was positively identified as not existing by the API.
            This key is used as a flag to distinguish 404s due to something not existing from other 404s such as mismatched routes or LB settings.
...
    BridgrSkillGapNode:
      type: object
      required:
        - uuid
        - id
        - graph_uuid
        - node_key
        - display_name
        - created_at
        - updated_at
      properties:
        uuid:
          type: string
          format: uuid
        id:
          type: integer
          format: int64
        graph_uuid:
          type: string
          format: uuid
        node_key:
          type: string
        display_name:
          type: string
        description:
          type: string
          nullable: true
        proficiency_hint:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        evidence:
          type: object
          additionalProperties: true
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        position_x:
          type: integer
          format: int32
          nullable: true
        position_y:
          type: integer
          format: int32
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BridgrSkillGapNodeListResponse:
      type: object
      required:
        - nodes
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/BridgrSkillGapNode'
...
    CreateBridgrSkillGapNodeRequest:
      type: object
      required:
        - node_key
        - display_name
      properties:
        node_key:
          type: string
        display_name:
          type: string
        description:
          type: string
          nullable: true
        proficiency_hint:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        evidence:
          type: object
          additionalProperties: true
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        position_x:
          type: integer
          format: int32
          nullable: true
        position_y:
          type: integer
          format: int32
          nullable: true
    BridgrSkillGapEdge:
      type: object
      required:
        - uuid
        - id
        - graph_uuid
        - from_node_uuid
        - to_node_uuid
        - relation
        - created_at
        - updated_at
      properties:
        uuid:
          type: string
          format: uuid
        id:
          type: integer
          format: int64
        graph_uuid:
          type: string
          format: uuid
        from_node_uuid:
          type: string
          format: uuid
        to_node_uuid:
          type: string
          format: uuid
        relation:
          type: string
        weight:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BridgrSkillGapEdgeListResponse:
      type: object
      required:
        - edges
      properties:
        edges:
          type: array
          items:
            $ref: '#/components/schemas/BridgrSkillGapEdge'
    CreateBridgrSkillGapEdgeRequest:
      type: object
      required:
        - from_node_uuid
        - to_node_uuid
        - relation
      properties:
        from_node_uuid:
          type: string
          format: uuid
        to_node_uuid:
          type: string
          format: uuid
        relation:
          type: string
        weight:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
...